[RENAME/#25] 패키지 이름 변경
