+++ feature/evaluate/src/main/java/com/my/version/feature/evaluate/record/state/EvaluationRecordUiState.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.evaluate.record.model
+package com.my.version.feature.evaluate.record.state
 
 import com.my.version.core.domain.entity.MusicAudioFile
